move generation module tests to cmapbuilder::grid
